.claude/skills/scout-api/scripts/requirements.txt
tikhub
openpyxl
python-dotenv
